test(task-migrate-clean-batch): 迁 face_number_overlay 的 12 个纯数学用例到 gui_unit_test
420.5 Step 11。部分迁移 12/15，拆分线就是作者自己留下的 ImGui category：
"unit" 的 12 个不需要 ImGui 上下文，"screenshot" 的 2 个要字体图集/真绘制，
再加 face_number_aggregate_from_topology_parity 走 BuildCrystalMeshData，三个留下。

- 新增 test/unit-correctness/gui/test_face_number_overlay.cpp（12 个 TEST）
  RotX180 / ReferenceFrontFacing 两个 helper 与那批 using 声明随之迁走；
  Identity4x4 两边都要，两边各留一份
- 原文件删 12 个块（15→3），头注释重写，随迁移空转的 helper 与 using 清掉

一处断言形态变更（唯一不是 1:1 的地方，故单独说明）：
face_number_normal_matches_center_frame 里的
`if (dot <= 0.0f) IM_ERRORF(...)` 改写为带 `<<` 诊断信息的 EXPECT_GT(dot, 0.0f)。
IM_ERRORF 是条件失败而非断言宏，机械计数器数不到它，所以断言总数记为 51→52；
判据、诊断字段与失败条件完全一致，只是从"手写 if 再报错"变成 gtest 断言。

核验：源文件侧断言减量 IM_CHECK 33 + IM_CHECK_EQ 9 + IM_CHECK_GT 4 + IM_CHECK_GE 2
+ IM_CHECK_LT 2 + IM_CHECK_LE 1 = 51（另 IM_ERRORF 1 处，见上），
新文件 EXPECT_TRUE 33 / EXPECT_EQ 9 / EXPECT_GT 5 / EXPECT_GE 2 / EXPECT_LT 2 /
EXPECT_LE 1 = 52，逐类对上；
gui_test 静态用例数 491→479；gui_unit_test 84→96 全绿。
